requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
Flask==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kafka-python==2.0.0
MarkupSafe==2.1.5
pymongo==4.8.0
requests==2.32.3
urllib3==2.2.2
Werkzeug==3.0.3
